Updated package description about HC30PL limitation
Added note about HC30PL limitation about the flange needing to point downward which cannot be accounted for by the URDF file. Increase B-axis motion so that compounded angles of the L,U and B axes can keep the flange model with 15 degrees of the ground.
Reverted package version to 0.1.0